VectorStoreRetriever Memory Module (#2791)
Adds a memory object that stores memories in a vecttor store via a
retriever interface.

Updates the VectorStoreRetriever interface to expose `add_documents()`
methods.

So far doesn't inherit from ChatMemory since the initial usage is likely
to be around storing intermediate steps. However it would be
straightforward to alter or add an addition memory class for that.

Adds a notebook.